+++ executor/invoker/src/commonMain/kotlin/io/github/charlietap/chasm/executor/invoker/instruction/memory/store/I64Store32Executor.kt
@@ -32,7 +32,7 @@ internal inline fun I64Store32Executor(
     val valueToStore = stack.popI64()
 
     val baseAddress = stack.popI32()
-    val effectiveAddress = baseAddress + instruction.memArg.offset.toInt()
+    val effectiveAddress = baseAddress + instruction.memArg.offset
 
     boundsChecker(effectiveAddress, 4, memory.size) {
         writer(memory.data, effectiveAddress, valueToStore).bind()
